move around code change package to flag-gen to avoid conflict with flag-generator binary test code generation
