+++ flag-gen/templates/altsrc_flags_generated.gotpl
