+++ flag-gen/templates/cli_flags_generated.gotpl
